requirements.txt
langgraph
langchain
chainlit
langchain_community
langchain-together
langchain-openai
langchain_anthropic
matplotlib
fastapi
uvicorn
mysql-connector-python
pandas
scikit-learn
PyPDF2